ci: fix fmt/clippy/test/MSRV failures
- Bump MSRV to 1.85 (noodles-core 0.19 requires edition2024).
- Bump workspace version to 0.1.1.
- cargo fmt + cargo clippy --fix sweep across workspace.
- Allow port-fidelity clippy lints workspace-wide (too_many_arguments,
  needless_range_loop, missing_safety_doc, doc_lazy_continuation,
  doc_overindented_list_items, branches_sharing_code, if_same_then_else,
  type_complexity); inherit via [lints] workspace = true in each crate.
- write! -> writeln! in star-core/error.rs; #[allow(result_unit_err)] on
  two C++-matching nucleotide helpers; drop redundant let rebinding and
  unnecessary pointer casts.
- .gitattributes: mark crates/star-params/src/parametersDefault as -text
  so Windows checkout keeps LF and the byte-length assertion passes.

+++ crates/star-align/src/mult_map_select.rs
@@ -79,7 +79,7 @@ impl ReadAlign {
         } else {
             let mut nbest: usize = 0;
             let u_sentinel = u64::MAX;
-            let out_sam_mult_nmax_set = (p.out_sam_multi_nmax as i64) != -1;
+            let out_sam_mult_nmax_set = p.out_sam_multi_nmax != -1;
             if p.out_multimapper_order_random || out_sam_mult_nmax_set {
                 let n_tr = self.n_tr as usize;
                 let max_score = self.max_score;
